Return EquinoxTimestamp as int

+++ src/Cmixin/BusinessDay/Calculator/HolidayCalculator.php
@@ -146,7 +146,7 @@ class HolidayCalculator extends CalculatorBase
         return date('m-d', $this->getEquinoxTimestamp($match));
     }
 
-    protected function getEquinoxTimestamp($match): float
+    protected function getEquinoxTimestamp($match): int
     {
         $month = strtolower($match[1]);
         $deltas = [
@@ -164,7 +164,7 @@ class HolidayCalculator extends CalculatorBase
         $hours = isset($match[3]) ? $match[3] * 1 : 0;
         $minutes = isset($match[4]) ? $match[4] * 1 : 0;
 
-        return round(
+        return (int) round(
             gmmktime(0, 0, 0, 1, 1, 2000) +
             (79.3125 + ($this->year - 2000) * 365.2425) * 86400 + $delta +
             ($hours * 60 + $minutes) * 60 * $sign
